[PPS | Refactor - 1] : Create dashboard client to stop using server component for user experience; Refactoring board APIs
