+++ package.json
@@ -48,6 +48,7 @@
     "react-dom": "^18",
     "react-hook-form": "^7.53.1",
     "react-icons": "^5.3.0",
+    "server-only": "^0.0.1",
     "sonner": "^1.5.0",
     "tailwind-merge": "^2.5.4",
     "tailwindcss-animate": "^1.0.7",
