+++ src/features/auth/utils.tsx
@@ -1,3 +1,5 @@
+import "server-only";
+
 import { redirect } from "next/navigation";
 
 import { auth } from "@/auth";
