+++ src/features/boards/api/use-create-board.ts
@@ -1,7 +1,10 @@
+import { useRouter } from "next/navigation";
+
 import { client } from "@/lib/hono";
+import { InferRequestType, InferResponseType } from "hono";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 
-import { InferRequestType, InferResponseType } from "hono";
+import { toast } from "sonner";
 
 type RequestType = InferRequestType<
   (typeof client.api.boards)["$post"]
@@ -10,6 +13,7 @@ type RequestType = InferRequestType<
 type ResponseType = InferResponseType<(typeof client.api.boards)["$post"], 200>;
 
 export const useCreateBoard = () => {
+  const router = useRouter();
   const queryClient = useQueryClient();
   const mutation = useMutation<ResponseType, Error, RequestType>({
     mutationFn: async (json) => {
@@ -21,8 +25,15 @@ export const useCreateBoard = () => {
 
       return await response.json();
     },
-    onSuccess: () => {
+    onSuccess: ({ data }) => {
+      toast.success(`New Board "${data.title}"`);
+
+      router.push(`/board/${data.id}`);
+
       queryClient.invalidateQueries({ queryKey: ["boards"] });
+    },
+    onError: () => {
+      toast.error("Failed to create board");
     },
   });
   return mutation;
