+++ src/features/boards/api/use-get-boards.ts
@@ -4,7 +4,7 @@ import { InferResponseType } from "hono";
 import { useQuery } from "@tanstack/react-query";
 
 export type ResponseType = InferResponseType<
-  (typeof client.api.boards)["boards-list"],
+  (typeof client.api.boards)["boards"],
   200
 >;
 
@@ -12,7 +12,7 @@ export const useGetBoards = () => {
   const query = useQuery({
     queryKey: ["boards"],
     queryFn: async () => {
-      const response = await client.api.boards["boards-list"].$get();
+      const response = await client.api.boards["boards"].$get();
 
       if (!response.ok) {
         throw new Error("Something went wrong!");
